chore: Exceptions, Typing and Business Logic changes.

+++ backend/src/use_cases/recipes/delete_recipe_use_case.py
@@ -1,39 +1,37 @@
 from uuid import UUID
 
-from fastapi import Depends, HTTPException, status
+from fastapi import Depends
 
 from ...repositories.recipe_repository import RecipeRepository
+from ...utils.exceptions import *
 
 
 class DeleteRecipeUseCase:
-    def __init__(self,
-                 repository: RecipeRepository = Depends(RecipeRepository),
-                 ) -> None:
+    def __init__(self, repository: RecipeRepository = Depends(RecipeRepository)) -> None:
         self._repository = repository
 
     def execute(
             self,
             token: dict[str, str],
-            recipe_id: UUID = None,
-            recipe_title: str = None
+            recipe_id: UUID | None = None,
+            recipe_title: str | None = None
     ) -> None:
-        if recipe_id is not None:
+        if recipe_id:
             recipe = self._repository.get_by_id(recipe_id)
-        elif recipe_title is not None:
+        elif recipe_title:
             recipe = self._repository.get_by_title(recipe_title)
         else:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="Recipe not found"
-            )
+            raise BlankRecipeIdAndTitle()
+
+        if not recipe:
+            raise RecipeNotFound()
 
         RECIPE_AUTHOR = recipe.user_id
-        CURRENT_USER_ID = token["user_id"]
+        CURRENT_USER_ID = UUID(token["id_user"])
+
+        print(RECIPE_AUTHOR, CURRENT_USER_ID)
 
         if RECIPE_AUTHOR != CURRENT_USER_ID:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail="You are not authorized to delete this recipe."
-            )
+            raise UnauthorizedRecipeDelete()
 
         self._repository.delete(recipe)
